// src/types.ts
export enum IPC {
	DISCORD_GET_INSTANCES = "DISCORD_GET_INSTANCES",
	DOWNLOADER_DOWNLOAD_MOD = "DOWNLOADER_DOWNLOAD_MOD",
	GET_PROFILES = "GET_PROFILES",
	GET_PROFILE = "GET_PROFILE",
	ADD_PROFILE = "ADD_PROFILE",
	UPDATE_PROFILE = "UPDATE_PROFILE",
	DELETE_PROFILE = "DELETE_PROFILE",
	START_PROFILE = "START_PROFILE",
	CREATE_DESKTOP_SHORTCUT = "CREATE_DESKTOP_SHORTCUT",
}